Fix warning

+++ source/application.cpp
@@ -611,8 +611,8 @@ bool MsgSleep(int aSleepDuration, MessageMode aMode)
 				HWND child_under_cursor = WindowFromPoint(msg.pt);
 				HWND parent_gui = GetParent(msg.hwnd);
 				if ( child_under_cursor != msg.hwnd && (pgui = GuiType::FindGui(parent_gui))
-					&& ((msg.message == WM_MOUSEWHEEL && pgui->mStyle & WS_VSCROLL && pgui->mVScroll->nPage <= pgui->mVScroll->nMax)
-					|| (msg.message == WM_MOUSEHWHEEL && pgui->mStyle & WS_HSCROLL && pgui->mHScroll->nPage <= pgui->mHScroll->nMax)))
+					&& ((msg.message == WM_MOUSEWHEEL && pgui->mStyle & WS_VSCROLL && (int)pgui->mVScroll->nPage <= pgui->mVScroll->nMax)
+					|| (msg.message == WM_MOUSEHWHEEL && pgui->mStyle & WS_HSCROLL && (int)pgui->mHScroll->nPage <= pgui->mHScroll->nMax)))
 				{
 					SendMessage(parent_gui, msg.message, msg.wParam, msg.lParam);
 					continue;
